utils module
Refactored downloading and file naming functions into conference.utils.py.
